feat: create delete implementation

+++ jnosql-mapping/jnosql-mapping-core/src/main/java/org/eclipse/jnosql/mapping/core/query/AnnotationOperation.java
@@ -105,7 +105,7 @@ public enum AnnotationOperation {
             } else if (returnType.isBoolean()) {
                 return true;
             } else if (returnType.isInt()) {
-                return StreamSupport.stream(entities.spliterator(), false).count();
+                return (int) StreamSupport.stream(entities.spliterator(), false).count();
             }
             return null;
         }
